requirements.txt
alembic==1.9.1
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
pycodestyle==2.10.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
